linux/requirements.txt
pynput==1.7.6
pystray==0.19.5
Pillow==10.0.0
